Changes to make release 0.11.0 compile on FreeBSD.
There were some issues with missing linker dependencies. Added them to
the Makefile.

On FreeBSD backtrace() returns a size_t and not an int. The compiler
threw an error on truncation from 32 bit uint to 64 bit uint.

+++ src/libponyrt/platform/ponyassert.c
@@ -33,12 +33,12 @@ void ponyint_assert_fail(const char* expr, const char* file, size_t line,
     func, expr);
 
   void* addrs[256];
-  int depth = backtrace(addrs, 256);
+  size_t depth = backtrace(addrs, 256);
   char** strings = backtrace_symbols(addrs, depth);
 
   fputs("Backtrace:\n", stderr);
 
-  for(int i = 0; i < depth; i++)
+  for(size_t i = 0; i < depth; i++)
     printf("  %s\n", strings[i]);
 
   free(strings);
